- first stabilization of the test with the new API
git-svn-id: https://svn.cern.ch/reps/lcgdm/davix/trunk@7439 4525493e-7705-40b1-a816-d608a930855b

+++ test/unit/neon/test_neon.cpp
@@ -14,7 +14,7 @@ static const char* list_urls[] = {
     "http://localhost:rrrrrrr/path" /* error*/
 };
 
-static const int list_port[]={
+static const size_t list_port[]={
     80,
     8080
 };
@@ -46,7 +46,7 @@ TEST(testNeon, testParsing){
     g_logger_set_globalfilter(G_LOG_LEVEL_MASK);
     std::string url, host, path, proto;
     unsigned long port;
-    for(int i=0; i < len_list; ++i){
+    for(size_t i=0; i < len_list; ++i){
         if(failure[i] == false){
             parse_http_neon_url(std::string(list_urls[i]), proto, host, path, &port);
             ASSERT_EQ(0, proto.compare(std::string(list_proto[i])));
